MAT-2273 and some refactoring
