+++ cypress/integration/BonnieFHIR/WebInterface/Dashboard/UploadDialog/ErrorHandling.spec.js
@@ -1,9 +1,8 @@
 import * as helper from '../../../../../support/helpers'
 import * as bonnieLogin from '../../../../../support/BonnieFHIR/BonnieLoginLogout'
-import * as dashboard from '../../../../../pom/BonnieFHIR/WI/Dashboard'
 import * as importMeasureDialog from '../../../../../pom/BonnieFHIR/WI/ImportMeasureDialog'
+import * as bonnieUploadMeasure from '../../../../../support/BonnieFHIR/BonnieUploadMeasure'
 
-let VsacApiKey = Cypress.env('VSAC_API_KEY')
 
 
 describe('Dashboard: Upload Dialog: Error handling', () => {
@@ -17,41 +16,25 @@ describe('Dashboard: Upload Dialog: Error handling', () => {
         bonnieLogin.logout()
 
     })
+    it('QDM Measure Package: Verify unable to upload package', () => {
 
-    it('Missing FHIR Version from Measure Library CQL(package json in the zip file)', () => {
-
-        helper.enabledWithTimeout(dashboard.uploadBtn)
-        cy.get(dashboard.uploadBtn).click()
-
-        //upload the file to the modal
-        helper.visibleWithTimeout(importMeasureDialog.importMeasureDialog)
-        const fileToUpload = "MissingFHIRDeclarationCQL.zip"
-        helper.enabledWithTimeout(importMeasureDialog.fileImportInput)
-
-        cy.get(importMeasureDialog.fileImportInput).attachFile(fileToUpload)
-
-        //wait for VSAC Key field to display for the user, and enter Key
-        helper.visibleWithTimeout(importMeasureDialog.vsacApiKeyTextBox)
-        helper.enabledWithTimeout(importMeasureDialog.vsacApiKeyTextBox)
-        helper.enterText(importMeasureDialog.vsacApiKeyTextBox, VsacApiKey)
-
-        //click load button to import the measure
-        helper.enabled(importMeasureDialog.importLoadBtn)
-        helper.click(importMeasureDialog.importLoadBtn)
+        bonnieUploadMeasure.UploadMeasureToBonnie('CMS105_v5_8_Artifacts.zip',false)
 
         helper.visibleWithTimeout(importMeasureDialog.errorDialog)
 
         //verify the error message
-        cy.get(importMeasureDialog.modalBody).contains('The measure could not be loaded.')
-        cy.get(importMeasureDialog.modalBody).contains('Bonnie has encountered an error while trying to load the measure.')
-        cy.get(importMeasureDialog.modalBody).contains('If the problem continues, please report the issue on the BONNIE Issue Tracker.')
+
+        cy.get(importMeasureDialog.errorDialog).invoke('text').then((text) => {
+            expect(text).to.include('The uploaded measure bundle does not contain the proper FHIR JSON file.' +
+              'Please re-package and re-export your FHIR based measure from the MAT and try again.')
+            expect(text).to.include('If the problem continues, please report the issue on the BONNIE Issue Tracker.')
+        })
 
         cy.get(importMeasureDialog.modalCloseBtn).click()
 
-        helper.notVisibleWithTimeout(importMeasureDialog.modalBody)
+        helper.notVisibleWithTimeout(importMeasureDialog.errorDialog)
 
-        cy.get(importMeasureDialog.measureNameDiv).should('not.contain','MissingFHIRDeclarationCQL')
+        cy.get(importMeasureDialog.measureNameDiv).should('not.contain','CMS105_v5_8_Artifacts')
 
     })
-
 })
