+++ cypress/integration/BonnieFHIR/WebInterface/Patient/EditPatient.spec.js
@@ -1,6 +1,5 @@
 import * as helper from '../../../../support/helpers'
 import * as bonnieLogin from '../../../../support/BonnieFHIR/BonnieLoginLogout'
-import * as homePage from '../../../../pom/BonnieFHIR/WI/Homepage'
 import * as measureDetailsPage from '../../../../pom/BonnieFHIR/WI/MeasureDetailsPage'
 import * as deletePatient from '../../../../support/BonnieFHIR/DeletePatient'
 import * as deleteMeasure from '../../../../support/BonnieFHIR/DeleteMeasure'
@@ -9,105 +8,69 @@ import * as bonnieUploadMeasure from '../../../../support/BonnieFHIR/BonnieUploa
 
 describe('Create and then Edit New Patient', () => {
 
-  let measureName = 'CMS111TestMeasureNK'
-  let measureFileToUpload = 'ContinuousVariableCMS111.zip'
+  let measureName = 'FHIRmeasureCMS347'
+  let measureFileToUpload = 'FHIRmeasureCMS347.zip'
+
+  let lastNameSuffix = new Date().getTime()
+  let distinctLastName = 'President' + lastNameSuffix
 
   before('Login', () => {
+
     bonnieLogin.login()
+    bonnieUploadMeasure.UploadMeasureToBonnie(measureFileToUpload,false)
+    measureDetailsPage.navigateToMeasureDetails(measureName)
+
   })
   after('Log Out', () => {
+
+    deleteMeasure.DeleteMeasure(measureName)
+    helper.visibleWithTimeout(measureDetailsPage.measurePageNavigationBtn)
     bonnieLogin.logout()
+
   })
 
   it('Create and edit patient', () => {
-    uploadTestMeasure()
-
-    navigateToMeasureDetails(measureName)
-
-    const lastNameSuffix = new Date().getTime()
-    const distinctLastName = 'President' + lastNameSuffix
 
     cy.get(measureDetailsPage.patientListing).then((patientListing) => {
       const initialPatientCount = parseInt(patientListing.text())
       cy.log('patient count was:' + initialPatientCount)
 
       measureDetailsPage.clickAddPatient()
-      enterPatientCharacteristics(distinctLastName)
+      testPatientPage.enterPatientCharacteristics(distinctLastName)
+
       testPatientPage.clickSavePatient()
-      verifyPatientAdded(initialPatientCount, distinctLastName)
+
+      testPatientPage.verifyPatientAdded(initialPatientCount, distinctLastName)
+
       measureDetailsPage.navigateToHomeMeasurePage()
-      navigateToMeasureDetails(measureName)
+
+      measureDetailsPage.navigateToMeasureDetails(measureName)
+
       editPatientCharacteristics(distinctLastName)
+
       testPatientPage.clickSavePatient()
-      measureDetailsPage.navigateToHomeMeasurePage()
-      navigateToMeasureDetails(measureName)
+
+      testPatientPage.verifyPatientAdded(initialPatientCount, distinctLastName)
+
       deletePatient.DeletePatient(distinctLastName)
-      verifyPatientRemoved(initialPatientCount)
+      deletePatient.VerifyPatientRemoved(initialPatientCount)
 
     })
 
+    measureDetailsPage.navigateToHomeMeasurePage()
     helper.visibleWithTimeout(measureDetailsPage.measurePageNavigationBtn)
 
-    deleteMeasure.DeleteMeasure(measureName)
-
-    helper.visibleWithTimeout(measureDetailsPage.measurePageNavigationBtn)
   })
-
-  function uploadTestMeasure () {
-    bonnieUploadMeasure.UploadMeasureToBonnie(measureFileToUpload)
-  }
-
-  function navigateToMeasureDetails (measureName) {
-    cy.log('navigateToMeasureDetails')
-    cy.get(homePage.measure).contains(measureName).click()
-    // cy.wait(1000)
-    cy.get(measureDetailsPage.measureDetailsTitle).should('contain.text', 'Measure details')
-    cy.log('navigateToMeasureDetails - done')
-  }
-
-  function enterPatientCharacteristics (lastName) {
-    cy.log('enterPatientCharacteristics')
-    cy.get(testPatientPage.lastNameTextField).type(lastName)
-    cy.get(testPatientPage.firstNameTextField).type('Current')
-
-    cy.get(testPatientPage.patientDescriptionTextField).type('Patient is very special')
-    cy.get(testPatientPage.dateofBithField).type('01/01/1950')
-    cy.get(testPatientPage.patientDescriptionTextField).click()
-    cy.get(testPatientPage.raceDropdown).select('Asian')
-    cy.get(testPatientPage.genderDropdown).select('Male')
-    cy.get(testPatientPage.ethnicityDropdown).select('Not Hispanic or Latino')
-    cy.log('enterPatientCharacteristics - done')
-  }
-
-  function getPatientRecord (lastName) {
-    return cy.get(measureDetailsPage.measureCalculationPanel).contains(lastName).parents(measureDetailsPage.patient)
-  }
-
-  function editPatientCharacteristics (lastName) {
-    cy.log('editPatientCharacteristics')
-    cy.get(measureDetailsPage.patientExpandBtn).click()
-    cy.get(measureDetailsPage.patientEditBtn).click()
-
-    cy.get(testPatientPage.lastNameTextField).clear().type(lastName)
-    cy.get(testPatientPage.firstNameTextField).clear().type('Future')
-
-    cy.get(testPatientPage.patientDescriptionTextField).clear().type('Description after edited on patient')
-    cy.get(testPatientPage.dateofBithField).type('01/01/1976')
-  }
-
-  function verifyPatientAdded (initialPatientCount, lastName) {
-    cy.log('verifyPatientAdded')
-    cy.get(measureDetailsPage.newStatus).should('have.text', 'NEW')
-    cy.get(measureDetailsPage.patientListing).should('have.text', (initialPatientCount + 1).toString())
-    getPatientRecord(lastName).find(measureDetailsPage.patientStatus).should('contain.text', 'pass')
-    cy.log('verifyPatientAdded - done')
-  }
-
-  function verifyPatientRemoved (initialPatientCount) {
-    cy.log('verifyPatientRemoved')
-    cy.get(measureDetailsPage.newStatus).should('have.text', 'NEW')
-    cy.get(measureDetailsPage.patientListing).should('have.text', (initialPatientCount).toString())
-    cy.log('verifyPatientRemoved - done')
-  }
-
 })
+
+function editPatientCharacteristics (lastName) {
+  cy.log('editPatientCharacteristics')
+  cy.get(measureDetailsPage.patientExpandBtn).click()
+  cy.get(measureDetailsPage.patientEditBtn).click()
+
+  cy.get(testPatientPage.lastNameTextField).clear().type(lastName)
+  cy.get(testPatientPage.firstNameTextField).clear().type('Future')
+
+  cy.get(testPatientPage.patientDescriptionTextField).clear().type('Description after edited on patient')
+  cy.get(testPatientPage.dateofBithField).type('01/01/1976')
+}
